Refactor: move transformer/model to nlp/transformer.
PiperOrigin-RevId: 286325224
