+++ official/nlp/transformer/model_utils_test.py
@@ -18,26 +18,18 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-import tensorflow as tf  # pylint: disable=g-bad-import-order
+import tensorflow as tf
 
-from official.transformer.model import model_utils
-from official.utils.misc import keras_utils
+from official.nlp.transformer import model_utils
 
 NEG_INF = -1e9
 
 
 class ModelUtilsTest(tf.test.TestCase):
 
-  def setUp(self):
-    super(ModelUtilsTest, self).setUp()
-    if keras_utils.is_v2_0:
-      tf.compat.v1.disable_eager_execution()
-
   def test_get_padding(self):
     x = tf.constant([[1, 0, 0, 0, 2], [3, 4, 0, 0, 0], [0, 5, 6, 0, 7]])
     padding = model_utils.get_padding(x, padding_value=0)
-    with self.session() as sess:
-      padding = sess.run(padding)
 
     self.assertAllEqual([[0, 1, 1, 1, 0], [0, 0, 1, 1, 1], [1, 0, 0, 1, 0]],
                         padding)
@@ -47,8 +39,6 @@ class ModelUtilsTest(tf.test.TestCase):
     bias = model_utils.get_padding_bias(x)
     bias_shape = tf.shape(bias)
     flattened_bias = tf.reshape(bias, [3, 5])
-    with self.session() as sess:
-      flattened_bias, bias_shape = sess.run((flattened_bias, bias_shape))
 
     self.assertAllEqual([[0, NEG_INF, NEG_INF, NEG_INF, 0],
                          [0, 0, NEG_INF, NEG_INF, NEG_INF],
@@ -59,8 +49,6 @@ class ModelUtilsTest(tf.test.TestCase):
   def test_get_decoder_self_attention_bias(self):
     length = 5
     bias = model_utils.get_decoder_self_attention_bias(length)
-    with self.session() as sess:
-      bias = sess.run(bias)
 
     self.assertAllEqual([[[[0, NEG_INF, NEG_INF, NEG_INF, NEG_INF],
                            [0, 0, NEG_INF, NEG_INF, NEG_INF],
@@ -71,4 +59,5 @@ class ModelUtilsTest(tf.test.TestCase):
 
 
 if __name__ == "__main__":
+  assert tf.version.VERSION.startswith('2.')
   tf.test.main()
