+++ official/r1/transformer/embedding_layer.py
@@ -20,7 +20,6 @@ from __future__ import print_function
 
 import tensorflow as tf  # pylint: disable=g-bad-import-order
 
-from official.transformer.model import model_utils
 from official.r1.utils import tpu as tpu_utils
