--- official/transformer/model/__init__.py
@@ -1,20 +1,0 @@
-# Copyright 2019 The TensorFlow Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-"""Bring in the shared legacy Transformer modules into this module."""
-
-from official.r1.transformer import transformer
-from official.r1.transformer import ffn_layer
-from official.r1.transformer import embedding_layer
-from official.r1.transformer import attention_layer
