+++ official/transformer/v2/transformer.py
@@ -23,7 +23,7 @@ from __future__ import print_function
 
 import tensorflow as tf
 
-from official.transformer.model import model_utils
+from official.nlp.transformer import model_utils
 from official.transformer.utils.tokenizer import EOS_ID
 from official.transformer.v2 import attention_layer
 from official.transformer.v2 import beam_search
