+++ official/transformer/v2/transformer_test.py
@@ -20,7 +20,7 @@ from __future__ import print_function
 
 import tensorflow as tf
 
-from official.transformer.model import model_params
+from official.nlp.transformer import model_params
 from official.transformer.v2 import transformer
